dont convert to int

+++ network_stats.py
@@ -300,21 +300,21 @@ def process_pkts(pktreader, output_cb, live, local_network_addresses, packet_sta
             if(key.ip1 == src and key.port1 == sport and key.ip2 == dst and
                key.port2 == dport and key.proto == prot):
                 if conn_bucket[key]['1to2Packets'] == 0:
-                    conn_bucket[key]['1to2StartTime'] = int((pktts + pktms/1e6)*1000)
+                    conn_bucket[key]['1to2StartTime'] = ((pktts + pktms/1e6)*1000)
                 conn_bucket[key]['1to2Bytes'] += ip_len
                 conn_bucket[key]['1to2Packets'] += 1
                 if packet_stats:
-                    conn_bucket[key]['1to2PacketTimes'] += str(conn_bucket[key]['1to2StartTime'] - int((pktts + pktms/1e6)*1000)) + ','
+                    conn_bucket[key]['1to2PacketTimes'] += str(conn_bucket[key]['1to2StartTime'] - ((pktts + pktms/1e6)*1000)) + ','
                     conn_bucket[key]['1to2PacketSize'] += str(ip_len) + ','
                     conn_bucket[key]['packet_dir'] += "1;"
             elif (key.ip2 == src and key.port2 == sport and key.ip1 == dst and
                   key.port1 == dport and key.proto == prot):
                 if conn_bucket[key]['2to1Packets'] == 0:
-                    conn_bucket[key]['2to1StartTime'] = int((pktts + pktms/1e6)*1000)
+                    conn_bucket[key]['2to1StartTime'] = ((pktts + pktms/1e6)*1000)
                 conn_bucket[key]['2to1Bytes'] += ip_len
                 conn_bucket[key]['2to1Packets'] += 1
                 if packet_stats:
-                    conn_bucket[key]['2to1PacketTimes'] += str(conn_bucket[key]['2to1StartTime'] - int((pktts + pktms/1e6)*1000)) + ','
+                    conn_bucket[key]['2to1PacketTimes'] += str(conn_bucket[key]['2to1StartTime'] - ((pktts + pktms/1e6)*1000)) + ','
                     conn_bucket[key]['2to1PacketSize'] += str(ip_len) + ','
                     conn_bucket[key]['packet_dir'] += "2;"
             else:
